Fixed unimpactful syntax errors
As there was no CYCLESPERSECOND defined in MotorMovement.py, it was complaining that it didn't exist but the code was running anyway as we define CYCLESPERSECOND in RobotArmControl.py. It is also a required variable if you want to be able to change the cycles per second for every motor.